--- tests/extract/huggingface/conftest.py
@@ -1,674 +1,0 @@
-# SPDX-FileContributor: Arthit Suriyawongkul
-# SPDX-FileCopyrightText: 2026-present Arthit Suriyawongkul
-# SPDX-FileType: SOURCE
-# SPDX-License-Identifier: Apache-2.0
-
-"""Tests for the Hugging Face model metadata extractor.
-
-Mock-patch helpers and per-model card/config fixtures used by the
-sibling ``test_huggingface_*.py`` modules live in the topic-grouped
-``_hf_patches_*`` submodules and are re-exported here.
-
-See also: _hf_patches_base.py, _hf_patches_text_generation_pretrained.py,
-_hf_patches_text_generation_instruct.py,
-_hf_patches_text_generation_regional.py, _hf_patches_gated_metadata.py,
-_hf_patches_speech_audio.py, _hf_patches_multimodal.py,
-_hf_patches_omni_modal.py, _hf_patches_embeddings.py, _hf_patches_vision.py,
-_hf_patches_structured_text.py, _hf_patches_generative_3d.py.
-"""
-
-from __future__ import annotations
-
-from ._hf_patches_base import (
-    _MISTRAL_CARD_DATA,
-    _MISTRAL_CONFIG,
-    _MISTRAL_TOKENIZER_CONFIG,
-    _make_card_data,
-    _patch_hf_calls,
-)
-from ._hf_patches_embeddings import (
-    _CLIP_JAPANESE_V2_CARD_DATA,
-    _CLIP_JAPANESE_V2_CONFIG,
-    _CODEBERTA_CARD_DATA,
-    _CODEBERTA_CONFIG,
-    _CROSS_ENCODER_CARD_DATA,
-    _CROSS_ENCODER_CONFIG,
-    _GTE_MODERNBERT_CARD_DATA,
-    _GTE_MODERNBERT_CONFIG,
-    _GTE_RERANKER_CARD_DATA,
-    _GTE_RERANKER_CONFIG,
-    _LINE_DISTILBERT_CARD_DATA,
-    _LINE_DISTILBERT_CONFIG,
-    _NOMIC_GGUF_CARD_DATA,
-    _RURI_CARD_DATA,
-    _RURI_CONFIG,
-    _SONOISA_CARD_DATA,
-    _SONOISA_CONFIG,
-    _WANGCHANX_LEGAL_CARD_DATA,
-    _WANGCHANX_LEGAL_CONFIG,
-    _patch_clip_japanese_v2,
-    _patch_codeberta,
-    _patch_cross_encoder,
-    _patch_dinov2,
-    _patch_distilbert_multilingual,
-    _patch_gabert,
-    _patch_granite_embed,
-    _patch_gte_modernbert,
-    _patch_gte_reranker,
-    _patch_legal_embed_ita,
-    _patch_line_distilbert,
-    _patch_nomic_gguf,
-    _patch_rad_dino,
-    _patch_ruri,
-    _patch_sonoisa,
-    _patch_timm_convnext,
-    _patch_uni2,
-    _patch_wangchanx_legal,
-    _patch_xlm_roberta_base,
-)
-from ._hf_patches_gated_metadata import (
-    _BERT_TURKISH_CARD_DATA,
-    _BERT_TURKISH_CONFIG,
-    _DEEPSEEK_CARD_DATA,
-    _DEEPSEEK_CONFIG,
-    _EUROLLM_1B7_CARD_DATA,
-    _EUROLLM_1B7_CONFIG,
-    _EUROLLM_TOKENIZER_SENTINEL,
-    _GEMMA_CARD_DATA,
-    _NLLB_CONFIG,
-    _NLLB_TOKENIZER_CONFIG,
-    _OPENEUROLLM_CARD_DATA,
-    _OPENEUROLLM_CONFIG,
-    _SEALLMS_CARD_DATA,
-    _SEALLMS_CONFIG,
-    _SERENGETI_CONFIG,
-    _SERENGETI_TOKENIZER_CONFIG,
-    _STANZA_DE_CARD_DATA,
-    _STANZA_FI_CARD_DATA,
-    _TALKIE_CARD_DATA,
-    _patch_aya_vision,
-    _patch_bert_turkish,
-    _patch_cohere_aya_23,
-    _patch_deepseek,
-    _patch_eurollm_1b7,
-    _patch_gemma,
-    _patch_hunyuan_mt,
-    _patch_hunyuan_mt7b,
-    _patch_ii_medical,
-    _patch_inkubalm,
-    _patch_mistral_medium,
-    _patch_nllb,
-    _patch_openeurollm,
-    _patch_opus_mt_th_en,
-    _patch_resnet18,
-    _patch_seallms,
-    _patch_serengeti,
-    _patch_stanza_de,
-    _patch_stanza_fi,
-    _patch_swin,
-    _patch_talkie,
-    _patch_wmt22_cometkiwi,
-)
-from ._hf_patches_generative_3d import (
-    _APPLE_SHARP_CARD_DATA,
-    _BLENDERLLM_CARD_DATA,
-    _BLENDERLLM_CONFIG,
-    _BLENDERLLM_GGUF_CARD_DATA,
-    _HY_MOTION_CARD_DATA,
-    _HY_MOTION_CONFIG,
-    _SHAP_E_CARD_DATA,
-    _STABLE_ZERO123_CARD_DATA,
-    _patch_apple_sharp,
-    _patch_blenderllm,
-    _patch_blenderllm_gguf,
-    _patch_ernie_image_turbo,
-    _patch_groot,
-    _patch_hy_motion,
-    _patch_openvla,
-    _patch_pi05,
-    _patch_shap_e,
-    _patch_stable_zero123,
-)
-from ._hf_patches_multimodal import (
-    _EXAONE45_33B_AWQ_CARD_DATA,
-    _EXAONE45_33B_AWQ_CONFIG,
-    _EXAONE45_33B_CARD_DATA,
-    _EXAONE45_33B_CONFIG,
-    _EXAONE45_33B_FP8_CARD_DATA,
-    _EXAONE45_33B_FP8_CONFIG,
-    _EXAONE45_33B_GGUF_CARD_DATA,
-    _KANANA_15V_CARD_DATA,
-    _KANANA_15V_CONFIG,
-    _KIMI_CARD_DATA,
-    _KIMI_CONFIG,
-    _SEALION_CARD_DATA,
-    _TIMELENS_CARD_DATA,
-    _TIMELENS_CONFIG,
-    _patch_arabic_legal_ocr,
-    _patch_blip_vqa,
-    _patch_boba_food_gguf,
-    _patch_deplot,
-    _patch_donut,
-    _patch_exaone45_33b,
-    _patch_exaone45_33b_awq,
-    _patch_exaone45_33b_fp8,
-    _patch_exaone45_33b_gguf,
-    _patch_jina_v4,
-    _patch_kanana_15v,
-    _patch_kimi,
-    _patch_layoutlm,
-    _patch_llava_video,
-    _patch_sealion_gguf,
-    _patch_sealion_vl,
-    _patch_timelens,
-    _patch_vilt_vqa,
-)
-from ._hf_patches_omni_modal import (
-    _AION_CARD_DATA,
-    _AION_CONFIG,
-    _BAGEL_CARD_DATA,
-    _BAGEL_CONFIG,
-    _LLADA2_MOE_CARD_DATA,
-    _LLADA2_MOE_CONFIG,
-    _MIMO_AUDIO_CARD_DATA,
-    _MIMO_AUDIO_CONFIG,
-    _MLX_GEMMA4_CARD_DATA,
-    _MLX_GEMMA4_CONFIG,
-    _MMADA_CARD_DATA,
-    _MMADA_CONFIG,
-    _ONNX_GEMMA4_CARD_DATA,
-    _ONNX_GEMMA4_CONFIG,
-    _SENSENOVA_CARD_DATA,
-    _SENSENOVA_CONFIG,
-    _patch_aion,
-    _patch_bagel,
-    _patch_llada2_moe,
-    _patch_mimo_audio,
-    _patch_mlx_gemma4,
-    _patch_mmada,
-    _patch_nemotron,
-    _patch_onnx_gemma4,
-    _patch_sensenova,
-)
-from ._hf_patches_speech_audio import (
-    _FIRERED_VAD_CARD_DATA,
-    _KOKORO_CARD_DATA,
-    _KOKORO_CONFIG,
-    _LLASA_3B_CARD_DATA,
-    _LLASA_3B_CONFIG,
-    _VOXTRAL_MINI_CARD_DATA,
-    _VOXTRAL_MINI_CONFIG,
-    _WAV2VEC2_JP_CARD_DATA,
-    _WAV2VEC2_JP_CONFIG,
-    _WAV2VEC2_JP_MODEL_INDEX,
-    _WHISPER_CARD_DATA,
-    _WHISPER_CONFIG,
-    _WHISPER_LANGUAGES,
-    _patch_firered_vad,
-    _patch_granite_speech,
-    _patch_indic_conformer,
-    _patch_ipa_whisper,
-    _patch_kokoro,
-    _patch_llasa_3b,
-    _patch_mimo_asr_gguf,
-    _patch_omnivoice,
-    _patch_omnivoice_bf16,
-    _patch_pyannote_diar,
-    _patch_seamless_m4t,
-    _patch_vibevoice_asr,
-    _patch_voxtral_mini,
-    _patch_wav2vec2_id_jv_su,
-    _patch_wav2vec2_jp,
-    _patch_whisper,
-)
-from ._hf_patches_structured_text import (
-    _MOIRAI_CARD_DATA,
-    _MOIRAI_CONFIG,
-    _SAP_RPT_CARD_DATA,
-    _SUGOI_CARD_DATA,
-    _TRADEPULSE_CARD_DATA,
-    _TRADEPULSE_CONFIG,
-    _VNTL_CARD_DATA,
-    _patch_aspect_finnlp_th,
-    _patch_deberta_human_value,
-    _patch_falconsai,
-    _patch_fineweb_edu,
-    _patch_hy_mt_gguf,
-    _patch_moirai,
-    _patch_privacy_filter,
-    _patch_protonx_legal,
-    _patch_sap_rpt,
-    _patch_sugoi,
-    _patch_tapas,
-    _patch_tradepulse,
-    _patch_vntl,
-)
-from ._hf_patches_text_generation_instruct import (
-    _BLOOMZ_7B1_CARD_DATA,
-    _BLOOMZ_7B1_CONFIG,
-    _GLM45_AIR_REAP_CARD_DATA,
-    _GLM45_AIR_REAP_CONFIG,
-    _MINIMAX_M2_CARD_DATA,
-    _MINIMAX_M2_CONFIG,
-    _OCCIGLOT_CARD_DATA,
-    _OCCIGLOT_CONFIG,
-    _OCCIGLOT_TOKENIZER_SENTINEL,
-    _OPENVINO_MIXTRAL_CARD_DATA,
-    _OPENVINO_MIXTRAL_CONFIG,
-    _PHARIA_ALIGNED_CARD_DATA,
-    _PHARIA_CONTROL_CARD_DATA,
-    _patch_bloomz_7b1,
-    _patch_crow_9b,
-    _patch_glm45_air_reap,
-    _patch_hermes_3_llama_3b,
-    _patch_llama_3_2_3b_instruct,
-    _patch_minimax_m2,
-    _patch_occiglot,
-    _patch_openvino_mixtral,
-    _patch_opt_iml,
-    _patch_pharia_aligned,
-    _patch_pharia_control,
-    _patch_qwen3_reap,
-    _patch_stablelm_zephyr,
-    _patch_tinyllama_chat,
-)
-from ._hf_patches_text_generation_pretrained import (
-    _BLOOM_CARD_DATA,
-    _BLOOM_CONFIG,
-    _LLAMA_CARD_DATA,
-    _OPENELM_270M_CARD_DATA,
-    _OPENELM_270M_CONFIG,
-    _QWEN3_235B_CARD_DATA,
-    _QWEN3_235B_CONFIG,
-    _QWEN3_235B_GENERATION_CONFIG,
-    _QWEN35_27B_CARD_DATA,
-    _QWEN35_27B_CONFIG,
-    _STARCODER2_CARD_DATA,
-    _STARCODER2_CONFIG,
-    _patch_bloom,
-    _patch_gpt_neo_2_7b,
-    _patch_granite_4_1_8b,
-    _patch_llama,
-    _patch_llama_3_2_3b,
-    _patch_openelm_270m,
-    _patch_opt_2_7b,
-    _patch_phi2,
-    _patch_qwen3_235b,
-    _patch_qwen35_27b,
-    _patch_starcoder2,
-)
-from ._hf_patches_text_generation_regional import (
-    _CHINDA_CARD_DATA,
-    _CHINDA_CONFIG,
-    _CHINDA_GGUF_CARD_DATA,
-    _FUJITSU_LLM_CARD_DATA,
-    _OPENTHAIGPT_CARD_DATA,
-    _OPENTHAIGPT_CONFIG,
-    _SAILOR2_20B_CARD_DATA,
-    _SAILOR2_20B_CONFIG,
-    _TILDEOPEN_30B_64K_CARD_DATA,
-    _TILDEOPEN_30B_64K_CONFIG,
-    _TILDEOPEN_30B_CARD_DATA,
-    _TILDEOPEN_30B_CONFIG,
-    _TILDEOPEN_TOKENIZER_SENTINEL,
-    _TYPHOON_CARD_DATA,
-    _TYPHOON_CONFIG,
-    _patch_chinda,
-    _patch_chinda_gguf,
-    _patch_darwin_kr_legal,
-    _patch_fujitsu_llm,
-    _patch_gpt_neox_jp,
-    _patch_laguna,
-    _patch_llmjp,
-    _patch_mallam,
-    _patch_openthaigpt,
-    _patch_qwen3_swallow,
-    _patch_sailor2_20b,
-    _patch_sealion_27b_it,
-    _patch_tildeopen_30b,
-    _patch_tildeopen_30b_64k,
-    _patch_typhoon,
-    _patch_wangchanglm,
-)
-from ._hf_patches_vision import (
-    _EXAONE_PATH_CARD_DATA,
-    _HRNETPOSE_CARD_DATA,
-    _LIGHTGLUE_CARD_DATA,
-    _LIGHTGLUE_CONFIG,
-    _RTDETR_COCO_CARD_DATA,
-    _RTDETR_COCO_O365_CARD_DATA,
-    _RTDETR_COCO_O365_CONFIG,
-    _WINDOWSEAT_CARD_DATA,
-    _patch_depth_pro,
-    _patch_exaone_path,
-    _patch_fibo_edit,
-    _patch_flood_image_detect,
-    _patch_granite_geo_flood,
-    _patch_hrnetpose,
-    _patch_laion_clip,
-    _patch_lightglue,
-    _patch_marigold,
-    _patch_rmbg14,
-    _patch_rmbg20,
-    _patch_rtdetr_coco,
-    _patch_rtdetr_coco_o365,
-    _patch_streetclip,
-    _patch_vitpose,
-    _patch_windowseat,
-)
-
-__all__ = [
-    "_AION_CARD_DATA",
-    "_AION_CONFIG",
-    "_APPLE_SHARP_CARD_DATA",
-    "_BAGEL_CARD_DATA",
-    "_BAGEL_CONFIG",
-    "_BERT_TURKISH_CARD_DATA",
-    "_BERT_TURKISH_CONFIG",
-    "_BLENDERLLM_CARD_DATA",
-    "_BLENDERLLM_CONFIG",
-    "_BLENDERLLM_GGUF_CARD_DATA",
-    "_BLOOMZ_7B1_CARD_DATA",
-    "_BLOOMZ_7B1_CONFIG",
-    "_BLOOM_CARD_DATA",
-    "_BLOOM_CONFIG",
-    "_CHINDA_CARD_DATA",
-    "_CHINDA_CONFIG",
-    "_CHINDA_GGUF_CARD_DATA",
-    "_CLIP_JAPANESE_V2_CARD_DATA",
-    "_CLIP_JAPANESE_V2_CONFIG",
-    "_CODEBERTA_CARD_DATA",
-    "_CODEBERTA_CONFIG",
-    "_CROSS_ENCODER_CARD_DATA",
-    "_CROSS_ENCODER_CONFIG",
-    "_DEEPSEEK_CARD_DATA",
-    "_DEEPSEEK_CONFIG",
-    "_EUROLLM_1B7_CARD_DATA",
-    "_EUROLLM_1B7_CONFIG",
-    "_EUROLLM_TOKENIZER_SENTINEL",
-    "_EXAONE45_33B_AWQ_CARD_DATA",
-    "_EXAONE45_33B_AWQ_CONFIG",
-    "_EXAONE45_33B_CARD_DATA",
-    "_EXAONE45_33B_CONFIG",
-    "_EXAONE45_33B_FP8_CARD_DATA",
-    "_EXAONE45_33B_FP8_CONFIG",
-    "_EXAONE45_33B_GGUF_CARD_DATA",
-    "_EXAONE_PATH_CARD_DATA",
-    "_FIRERED_VAD_CARD_DATA",
-    "_FUJITSU_LLM_CARD_DATA",
-    "_GEMMA_CARD_DATA",
-    "_GLM45_AIR_REAP_CARD_DATA",
-    "_GLM45_AIR_REAP_CONFIG",
-    "_GTE_MODERNBERT_CARD_DATA",
-    "_GTE_MODERNBERT_CONFIG",
-    "_GTE_RERANKER_CARD_DATA",
-    "_GTE_RERANKER_CONFIG",
-    "_HRNETPOSE_CARD_DATA",
-    "_HY_MOTION_CARD_DATA",
-    "_HY_MOTION_CONFIG",
-    "_KANANA_15V_CARD_DATA",
-    "_KANANA_15V_CONFIG",
-    "_KIMI_CARD_DATA",
-    "_KIMI_CONFIG",
-    "_KOKORO_CARD_DATA",
-    "_KOKORO_CONFIG",
-    "_LIGHTGLUE_CARD_DATA",
-    "_LIGHTGLUE_CONFIG",
-    "_LINE_DISTILBERT_CARD_DATA",
-    "_LINE_DISTILBERT_CONFIG",
-    "_LLADA2_MOE_CARD_DATA",
-    "_LLADA2_MOE_CONFIG",
-    "_LLAMA_CARD_DATA",
-    "_LLASA_3B_CARD_DATA",
-    "_LLASA_3B_CONFIG",
-    "_MIMO_AUDIO_CARD_DATA",
-    "_MIMO_AUDIO_CONFIG",
-    "_MINIMAX_M2_CARD_DATA",
-    "_MINIMAX_M2_CONFIG",
-    "_MISTRAL_CARD_DATA",
-    "_MISTRAL_CONFIG",
-    "_MISTRAL_TOKENIZER_CONFIG",
-    "_MLX_GEMMA4_CARD_DATA",
-    "_MLX_GEMMA4_CONFIG",
-    "_MMADA_CARD_DATA",
-    "_MMADA_CONFIG",
-    "_MOIRAI_CARD_DATA",
-    "_MOIRAI_CONFIG",
-    "_NLLB_CONFIG",
-    "_NLLB_TOKENIZER_CONFIG",
-    "_NOMIC_GGUF_CARD_DATA",
-    "_OCCIGLOT_CARD_DATA",
-    "_OCCIGLOT_CONFIG",
-    "_OCCIGLOT_TOKENIZER_SENTINEL",
-    "_ONNX_GEMMA4_CARD_DATA",
-    "_ONNX_GEMMA4_CONFIG",
-    "_OPENELM_270M_CARD_DATA",
-    "_OPENELM_270M_CONFIG",
-    "_OPENEUROLLM_CARD_DATA",
-    "_OPENEUROLLM_CONFIG",
-    "_OPENTHAIGPT_CARD_DATA",
-    "_OPENTHAIGPT_CONFIG",
-    "_OPENVINO_MIXTRAL_CARD_DATA",
-    "_OPENVINO_MIXTRAL_CONFIG",
-    "_PHARIA_ALIGNED_CARD_DATA",
-    "_PHARIA_CONTROL_CARD_DATA",
-    "_QWEN35_27B_CARD_DATA",
-    "_QWEN35_27B_CONFIG",
-    "_QWEN3_235B_CARD_DATA",
-    "_QWEN3_235B_CONFIG",
-    "_QWEN3_235B_GENERATION_CONFIG",
-    "_RTDETR_COCO_CARD_DATA",
-    "_RTDETR_COCO_O365_CARD_DATA",
-    "_RTDETR_COCO_O365_CONFIG",
-    "_RURI_CARD_DATA",
-    "_RURI_CONFIG",
-    "_SAILOR2_20B_CARD_DATA",
-    "_SAILOR2_20B_CONFIG",
-    "_SAP_RPT_CARD_DATA",
-    "_SEALION_CARD_DATA",
-    "_SEALLMS_CARD_DATA",
-    "_SEALLMS_CONFIG",
-    "_SENSENOVA_CARD_DATA",
-    "_SENSENOVA_CONFIG",
-    "_SERENGETI_CONFIG",
-    "_SERENGETI_TOKENIZER_CONFIG",
-    "_SHAP_E_CARD_DATA",
-    "_SONOISA_CARD_DATA",
-    "_SONOISA_CONFIG",
-    "_STABLE_ZERO123_CARD_DATA",
-    "_STANZA_DE_CARD_DATA",
-    "_STANZA_FI_CARD_DATA",
-    "_STARCODER2_CARD_DATA",
-    "_STARCODER2_CONFIG",
-    "_SUGOI_CARD_DATA",
-    "_TALKIE_CARD_DATA",
-    "_TILDEOPEN_30B_64K_CARD_DATA",
-    "_TILDEOPEN_30B_64K_CONFIG",
-    "_TILDEOPEN_30B_CARD_DATA",
-    "_TILDEOPEN_30B_CONFIG",
-    "_TILDEOPEN_TOKENIZER_SENTINEL",
-    "_TIMELENS_CARD_DATA",
-    "_TIMELENS_CONFIG",
-    "_TRADEPULSE_CARD_DATA",
-    "_TRADEPULSE_CONFIG",
-    "_TYPHOON_CARD_DATA",
-    "_TYPHOON_CONFIG",
-    "_VNTL_CARD_DATA",
-    "_VOXTRAL_MINI_CARD_DATA",
-    "_VOXTRAL_MINI_CONFIG",
-    "_WANGCHANX_LEGAL_CARD_DATA",
-    "_WANGCHANX_LEGAL_CONFIG",
-    "_WAV2VEC2_JP_CARD_DATA",
-    "_WAV2VEC2_JP_CONFIG",
-    "_WAV2VEC2_JP_MODEL_INDEX",
-    "_WHISPER_CARD_DATA",
-    "_WHISPER_CONFIG",
-    "_WHISPER_LANGUAGES",
-    "_WINDOWSEAT_CARD_DATA",
-    "_make_card_data",
-    "_patch_aion",
-    "_patch_apple_sharp",
-    "_patch_arabic_legal_ocr",
-    "_patch_aspect_finnlp_th",
-    "_patch_aya_vision",
-    "_patch_bagel",
-    "_patch_bert_turkish",
-    "_patch_blenderllm",
-    "_patch_blenderllm_gguf",
-    "_patch_blip_vqa",
-    "_patch_bloom",
-    "_patch_bloomz_7b1",
-    "_patch_boba_food_gguf",
-    "_patch_chinda",
-    "_patch_chinda_gguf",
-    "_patch_clip_japanese_v2",
-    "_patch_codeberta",
-    "_patch_cohere_aya_23",
-    "_patch_cross_encoder",
-    "_patch_crow_9b",
-    "_patch_darwin_kr_legal",
-    "_patch_deberta_human_value",
-    "_patch_deepseek",
-    "_patch_deplot",
-    "_patch_depth_pro",
-    "_patch_dinov2",
-    "_patch_distilbert_multilingual",
-    "_patch_donut",
-    "_patch_ernie_image_turbo",
-    "_patch_eurollm_1b7",
-    "_patch_exaone45_33b",
-    "_patch_exaone45_33b_awq",
-    "_patch_exaone45_33b_fp8",
-    "_patch_exaone45_33b_gguf",
-    "_patch_exaone_path",
-    "_patch_falconsai",
-    "_patch_fibo_edit",
-    "_patch_fineweb_edu",
-    "_patch_firered_vad",
-    "_patch_flood_image_detect",
-    "_patch_fujitsu_llm",
-    "_patch_gabert",
-    "_patch_gemma",
-    "_patch_glm45_air_reap",
-    "_patch_gpt_neo_2_7b",
-    "_patch_gpt_neox_jp",
-    "_patch_granite_4_1_8b",
-    "_patch_granite_embed",
-    "_patch_granite_geo_flood",
-    "_patch_granite_speech",
-    "_patch_groot",
-    "_patch_gte_modernbert",
-    "_patch_gte_reranker",
-    "_patch_hermes_3_llama_3b",
-    "_patch_hf_calls",
-    "_patch_hrnetpose",
-    "_patch_hunyuan_mt",
-    "_patch_hunyuan_mt7b",
-    "_patch_hy_motion",
-    "_patch_hy_mt_gguf",
-    "_patch_ii_medical",
-    "_patch_indic_conformer",
-    "_patch_inkubalm",
-    "_patch_ipa_whisper",
-    "_patch_jina_v4",
-    "_patch_kanana_15v",
-    "_patch_kimi",
-    "_patch_kokoro",
-    "_patch_laguna",
-    "_patch_laion_clip",
-    "_patch_layoutlm",
-    "_patch_legal_embed_ita",
-    "_patch_lightglue",
-    "_patch_line_distilbert",
-    "_patch_llada2_moe",
-    "_patch_llama",
-    "_patch_llama_3_2_3b",
-    "_patch_llama_3_2_3b_instruct",
-    "_patch_llasa_3b",
-    "_patch_llava_video",
-    "_patch_llmjp",
-    "_patch_mallam",
-    "_patch_marigold",
-    "_patch_mimo_asr_gguf",
-    "_patch_mimo_audio",
-    "_patch_minimax_m2",
-    "_patch_mistral_medium",
-    "_patch_mlx_gemma4",
-    "_patch_mmada",
-    "_patch_moirai",
-    "_patch_nemotron",
-    "_patch_nllb",
-    "_patch_nomic_gguf",
-    "_patch_occiglot",
-    "_patch_omnivoice",
-    "_patch_omnivoice_bf16",
-    "_patch_onnx_gemma4",
-    "_patch_openelm_270m",
-    "_patch_openeurollm",
-    "_patch_openthaigpt",
-    "_patch_openvino_mixtral",
-    "_patch_openvla",
-    "_patch_opt_2_7b",
-    "_patch_opt_iml",
-    "_patch_opus_mt_th_en",
-    "_patch_pharia_aligned",
-    "_patch_pharia_control",
-    "_patch_phi2",
-    "_patch_pi05",
-    "_patch_privacy_filter",
-    "_patch_protonx_legal",
-    "_patch_pyannote_diar",
-    "_patch_qwen35_27b",
-    "_patch_qwen3_235b",
-    "_patch_qwen3_reap",
-    "_patch_qwen3_swallow",
-    "_patch_rad_dino",
-    "_patch_resnet18",
-    "_patch_rmbg14",
-    "_patch_rmbg20",
-    "_patch_rtdetr_coco",
-    "_patch_rtdetr_coco_o365",
-    "_patch_ruri",
-    "_patch_sailor2_20b",
-    "_patch_sap_rpt",
-    "_patch_sealion_27b_it",
-    "_patch_sealion_gguf",
-    "_patch_sealion_vl",
-    "_patch_seallms",
-    "_patch_seamless_m4t",
-    "_patch_sensenova",
-    "_patch_serengeti",
-    "_patch_shap_e",
-    "_patch_sonoisa",
-    "_patch_stable_zero123",
-    "_patch_stablelm_zephyr",
-    "_patch_stanza_de",
-    "_patch_stanza_fi",
-    "_patch_starcoder2",
-    "_patch_streetclip",
-    "_patch_sugoi",
-    "_patch_swin",
-    "_patch_talkie",
-    "_patch_tapas",
-    "_patch_tildeopen_30b",
-    "_patch_tildeopen_30b_64k",
-    "_patch_timelens",
-    "_patch_timm_convnext",
-    "_patch_tinyllama_chat",
-    "_patch_tradepulse",
-    "_patch_typhoon",
-    "_patch_uni2",
-    "_patch_vibevoice_asr",
-    "_patch_vilt_vqa",
-    "_patch_vitpose",
-    "_patch_vntl",
-    "_patch_voxtral_mini",
-    "_patch_wangchanglm",
-    "_patch_wangchanx_legal",
-    "_patch_wav2vec2_id_jv_su",
-    "_patch_wav2vec2_jp",
-    "_patch_whisper",
-    "_patch_windowseat",
-    "_patch_wmt22_cometkiwi",
-    "_patch_xlm_roberta_base",
-]
